+++ Study/JetpackCompose_practice/JetpackComposeCloneCoding/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:allowBackup="true"
@@ -11,12 +11,17 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.JetpackComposeCloneCoding"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
-            android:name=".MVVM.NavActivity"
+            android:name=".MainActivity"
+            android:exported="false"
+            android:label="@string/title_activity_main"
+            android:theme="@style/Theme.JetpackComposeCloneCoding" />
+        <activity
+            android:name=".myMVVMPractice.NavActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.JetpackComposeCloneCoding">
+            android:theme="@style/Theme.JetpackComposeCloneCoding" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
