+++ Study/JetpackCompose_practice/JetpackComposeCloneCoding/app/src/main/java/com/jetpackPractice/jetpackcomposeclonecoding/ui/theme/Color.kt
@@ -9,3 +9,21 @@ val Pink80 = Color(0xFFEFB8C8)
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
+
+
+// 아래 부분이 추가한 색상들로 이렇게 만들면 상수로써 사용가능
+
+val Lilac = Color(0xFFCCB6DC)
+val Yellow = Color(0xFFFFCB66)
+val YellowVariant = Color(0xFFFFDE9F)
+val Coral = Color(0xFFF3A397)
+val White = Color(0xFFFFFFFF)
+val MintGreen = Color(0xFFACD6B8)
+val DarkGray = Color(0xFF2B2B2D)
+
+val DarkCoral = Color(0xFFF7A374)
+val DarkYellow = Color(0xFFFFCE6F)
+val Yellow_Awake = Color(0xFFffeac1)
+val Yellow_Rem = Color(0xFFffdd9a)
+val Yellow_Light = Color(0xFFffcb66)
+val Yellow_Deep = Color(0xFFff973c)
