+++ Study/JetpackCompose_practice/JetpackComposeCloneCoding/app/src/main/java/com/jetpackPractice/jetpackcomposeclonecoding/ui/theme/Type.kt
@@ -2,14 +2,21 @@ package com.jetpackPractice.jetpackcomposeclonecoding.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
+import com.jetpackPractice.jetpackcomposeclonecoding.R
+
+
+val fontFamily = FontFamily(
+    Font(R.font.maruburiregular)
+)
 
 // Set of Material typography styles to start with
 val Typography = Typography(
     bodyLarge = TextStyle(
-        fontFamily = FontFamily.Default,
+        fontFamily = fontFamily,
         fontWeight = FontWeight.Normal,
         fontSize = 16.sp,
         lineHeight = 24.sp,
@@ -31,4 +38,25 @@ val Typography = Typography(
         letterSpacing = 0.5.sp
     )
     */
+)
+
+val HeadingStyle = TextStyle(
+    fontSize = 24.sp,
+    fontWeight = FontWeight(600),
+    letterSpacing = 0.5.sp,
+    fontFamily = fontFamily
+)
+
+val SmallHeadingStyle = TextStyle(
+    fontSize = 16.sp,
+    fontWeight = FontWeight(600),
+    letterSpacing = 0.5.sp,
+    fontFamily = fontFamily
+)
+
+val LegendHeadingStyle = TextStyle(
+    fontSize = 10.sp,
+    fontWeight = FontWeight(600),
+    letterSpacing = 0.5.sp,
+    fontFamily = fontFamily
 )
