+++ Study/JetpackCompose_practice/JetpackComposeCloneCoding/app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">JetpackComposeCloneCoding</string>
     <string name="title_activity_nav">NavActivity</string>
+    <string name="title_activity_main">MainActivity</string>
 </resources>
